removed lazyload from components to Image component
